Modified Resource to not always require Send
It is possible that some tasks will need a resource that is
initialized at runtime, but is not send (for example, the motivating
example is a smoltcp interface). This patch makes it so that a
resource that is only ever used in one task does not need to be Send.

Compatibility issues:
 * Anything that depends on `<T as Resource>::Data : Send` should
   instead assert that `T: SharedResource`