Add the concept of local tasks with an example
Add test for spawning a local task from a different prio

This ensures that there is no method on the spawning
task's local spawner for spawning the local task
with the other priority.

Add test for spawning a local task from init

This ensures that there is no method in the globaly
available module with the same name as the local task.

Add passing test for local task with non Send/Sync arg

Ensure it is possible to pass a non send and non Sync
argument when spawning a local task. This is ok
since it will only be possible to spawn the task
on the same executor and priority level

Fix spawn-local-no-send-sync

Rename `is_local_task` to just `local_task` and allow omitting the value

Move example and remove success test

Update tests for - Rename `is_local_task` to just `local_task` and allow omitting the value

Move tests to rtic/ui

Revert fmt

Add exit to example

Fix example

Fix example

Update changelog